binary symmetric channel added

+++ include/encoder.h
@@ -9,7 +9,7 @@
 constexpr uint8_t SCALE = 64;
 
 // Signaturen 
-void compute_parity(int (&base)[ROWS][COLS], std::array<uint64_t,4> message, int8_t scale, std::array<uint64_t,4> &parity);
+void compute_parity(int8_t (&base)[ROWS][COLS], std::array<uint64_t,4> message, int8_t scale, std::array<uint64_t,4> &parity);
 
 
 #endif
